fix (kubernetes-itests) : Add Kubernetes 1.25.0 to E2E test suite matrix
Run integration tests on Kubernetes 1.25.0 as well
