+++ kubernetes-itests/src/test/resources/pod-evict-it.yml
@@ -1,0 +1,28 @@
+#
+# Copyright (C) 2015 Red Hat, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+---
+apiVersion: v1
+kind: Pod
+metadata:
+  name: pod-standard
+  labels:
+    pdb-scope: get
+spec:
+  containers:
+    - name: busybox
+      image: busybox
+      command: ["sleep", "36000"]
